--- testCode/testCode/testCode/Base/App/Game/RPG/MVC/Controllers/Main/RPGMainViewController.swift
@@ -1,22 +1,0 @@
-//
-//  RPGMainViewController.swift
-//  testCode
-//
-//  Created by YouOhshima on 2019/06/06.
-//  Copyright © 2019年 Shinkai Asari. All rights reserved.
-//
-
-import UIKit
-
-@IBDesignable class RPGMainViewController: UIViewController {
-    @IBOutlet weak var MainImageView: UIImageView!
-    @IBOutlet weak var gameStartButton: UIButton!
-    @IBOutlet weak var gameSettingButton: UIButton!
-    @IBOutlet weak var endButton: UIButton!
-    
-    @IBOutlet weak var MainTitleLabel: UILabel!
-    @IBOutlet weak var copyrightLabel: UILabel!
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-}
