--- testCode/testCode/testCode/Base/App/Game/RPG/MVC/Controllers/init/RPGStartViewController.swift
@@ -1,16 +1,0 @@
-//
-//  RPGStartViewController.swift
-//  testCode
-//
-//  Created by YouOhshima on 2019/06/06.
-//  Copyright © 2019年 Shinkai Asari. All rights reserved.
-//
-
-import UIKit
-
-@IBDesignable class RPGStartViewController: UIViewController {
-    @IBOutlet weak var StartImageView: UIImageView!
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-}
